requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.37.0
pydantic==2.12.0
python-dotenv==1.1.0
